% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrainbows.R
\name{inrainbows}
\alias{inrainbows}
\title{Radiohead Album Palette}
\usage{
inrainbows(album, n)
}
\arguments{
\item{n}{Number of colours required}

\item{option}{Album name to use for colour scale.}
}
\value{
Character vector of colour strings
}
\description{
Radiohead Album Palette
}
\examples{
inrainbows("kid_a", 6)
inrainbows("httf", 6)
}
